Refactor as ECS
Refactor the game code as more ECS-oriented, partitioning it in more
coherent systems for the inventory management, the level loading, and
the deserialization of the levels from disk.

Move levels definition to `assets/levels.json`, and load it via a custom
`TextAsset` asset. This ensures it can be loaded in wasm. Handle async
loading with "Loading..." text on main menu, and extra `MainMenuLoading`
app state.