Add boot screen with progress bar
Add a dedicated boot screen before the main menu, with a progress bar
showing the loading progress of the critical boot assets. This is
currently extremely fast on non-web platform (only 3 assets) therefore
is nearly invisible, but is hopefully future-proof. The progress bar is
animated via a material with hard-coded shaders, to not require any
shader loading.