Prevent cursor action outside play
Prevent any action with the cursor, and especially spawning anything, while the
cursor is disabled. Disable the cursor during the (TODO: so far invisible)
pre-game and post-game sequences, to avoid spawning an item before the game can
listen to victory check events, which otherwise can leave the game board stuck
in a victory state without the game module checking for it.